src/main/webapp/app/entities/queue/delete/queue-delete-dialog.component.ts
import { Component } from '@angular/core';
import { FormsModule } from '@angular/forms';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import SharedModule from 'app/shared/shared.module';
import { IQueue } from '../queue.model';
import { QueueService } from '../service/queue.service';
import { ITEM_DELETED_EVENT } from 'app/config/navigation.constants';

@Component({
  standalone: true,
  templateUrl: './queue-delete-dialog.component.html',
  imports: [SharedModule, FormsModule],
})
export class QueueDeleteDialogComponent {
  queue?: IQueue;

  constructor(protected queueService: QueueService, protected activeModal: NgbActiveModal) {}

  cancel(): void {
    this.activeModal.dismiss();
  }

  confirmDelete(id: number): void {
    this.queueService.delete(id).subscribe(() => {
      this.activeModal.close(ITEM_DELETED_EVENT);
    });
  }
}
